Platform: Use SMBIOS RAM frequency when possible
If Clover does not trust the SMBIOS RAM data for any reason, it will
use the SPD RAM frequency, even if it is slower (overclocking is only
reflected in SMBIOS). Record the maximum sane SMBIOS RAM frequency and
use it if it is greater than the SPD RAM frequency.

+++ rEFIt_UEFI/Platform/Platform.h
@@ -1293,7 +1293,7 @@ typedef struct {
 
 typedef struct {
 
-  UINT64        Frequency;
+  UINT32        Frequency;
   UINT32        Divider;
   UINT8         TRC;
   UINT8         TRP;
